Refactor Bench into Codec, under its own module
